lsp-intellij: Setup IntelliJ integration with LSP
An interesting package for allowing Emacs to do be a
little smarter with Kotlin code.  This however adds
some additional packages as depencies:
- company-lsp
- lsp-mode
- lsp-ui
- yasnippet

Current configuration is just the very basic.

One thing to note is lsp-intellij currently has port
8080 hard-coded to connect to the LSP server, so
there might be issues if there are other services
running that use the same port.